* Moved `mediaEntry` from `PKPlugin` to player. * Added `BasePlugin` as a base to all plugins, very minimalistic has logs for all PKPlugins methods and holds player reference. * Small fixes and clean ups
