+++ Plugins/Phoenix/PhoenixAnalyticsPlugin.swift
@@ -38,7 +38,7 @@ public class PhoenixAnalyticsPlugin: BaseOTTAnalyticsPlugin {
             parterId = pId
         }
         
-        guard let mediaEntry = self.mediaEntry else {
+        guard let mediaEntry = self.player.mediaEntry else {
             PKLog.error("send analytics failed due to nil mediaEntry")
             self.messageBus.post(PlayerEvent.PluginError(nsError: AnalyticsError.missingMediaEntry.asNSError))
             return nil
