+++ Plugins/Phoenix/TVPAPIAnalyticsPlugin.swift
@@ -27,7 +27,7 @@ public class TVPAPIAnalyticsPlugin: BaseOTTAnalyticsPlugin {
             return nil
         }
         
-        guard let mediaEntry = self.mediaEntry else {
+        guard let mediaEntry = self.player.mediaEntry else {
             PKLog.error("send analytics failed due to nil mediaEntry")
             self.messageBus.post(PlayerEvent.PluginError(nsError: AnalyticsError.missingMediaEntry.asNSError))
             return nil
